Fixed issues and added new features
