+++ CPPProject/x64/Debug/CPPProject.log
@@ -1,8 +1,15 @@
 ﻿  GuiManager.cpp
 C:\Program Files (x86)\Windows Kits\10\Include\10.0.22621.0\shared\minwindef.h(130,9): warning C4005: 'APIENTRY': macro redefinition
-D:\CPP Project\CPPProject\GuiManager.cpp(93,49): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
-D:\CPP Project\CPPProject\GuiManager.cpp(93,36): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
-D:\CPP Project\CPPProject\GuiManager.cpp(306,31): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(41,32): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(41,25): warning C4244: 'argument': conversion from 'int' to 'float', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(98,49): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(98,36): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(201,49): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(201,36): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(236,59): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(236,45): warning C4244: 'argument': conversion from 'float' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(400,31): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
+D:\CPP Project\CPPProject\src\GuiManager.cpp(438,31): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data
   CPPProject.vcxproj -> D:\CPP Project\CPPProject\x64\Debug\CPPProject.exe
   'pwsh.exe' is not recognized as an internal or external command,
   operable program or batch file.
