+++ CPPProject/x64/Debug/CPPProject.tlog/Cl.items.tlog
@@ -1,6 +1,6 @@
-D:\CPP Project\CPPProject\GuiManager.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\GuiManager.obj
 D:\CPP Project\CPPProject\src\CPPProject.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\CPPProject.obj
 D:\CPP Project\CPPProject\src\DataManager.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\DataManager.obj
+D:\CPP Project\CPPProject\src\GuiManager.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\GuiManager.obj
 D:\CPP Project\CPPProject\third_party\imgui-1.90.8\backends\imgui_impl_glfw.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\imgui_impl_glfw.obj
 D:\CPP Project\CPPProject\third_party\imgui-1.90.8\backends\imgui_impl_opengl3.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\imgui_impl_opengl3.obj
 D:\CPP Project\CPPProject\third_party\imgui-1.90.8\imgui.cpp;D:\CPP Project\CPPProject\CPPProject\x64\Debug\imgui.obj
